Move shared_utils and server_utils to chat_service/core/ (#2123)
* Move shared_utils and server_utils to chat_service/core/

* change imports more places
