+++ parlai/chat_service/services/messenger/messenger_manager.py
@@ -14,11 +14,11 @@ import logging
 import os
 
 from parlai.core.agents import create_agent
+import parlai.chat_service.core.server_utils as server_utils
+import parlai.chat_service.core.shared_utils as shared_utils
 from parlai.chat_service.services.messenger.agents import MessengerAgent
 from parlai.chat_service.core.manager_utils import ChatServiceMessageSocket
 from parlai.chat_service.services.messenger.message_sender import MessageSender
-import parlai.chat_service.services.messenger.shared_utils as shared_utils
-import parlai.chat_service.services.messenger.server_utils as server_utils
 from parlai.chat_service.core.chat_service_manager import ChatServiceManager
 
 parent_dir = os.path.dirname(os.path.abspath(__file__))
